微服务数据库脚本/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
colorama==0.4.6
